Convert CTRL to multistep on C20_WMDE_Test_02
