+++ .eslintrc.json
@@ -33,6 +33,7 @@
 			"code": 180
 		}],
 		"max-statements-per-line": [ "error", { "max": 2 } ],
+		"prefer-regex-literals": [ "off" ],
 		"react/jsx-uses-react": "error",
 		"react/jsx-uses-vars": "error",
 		"react/jsx-pascal-case": "error",
