+++ desktop/banner_ctrl.js
@@ -9,8 +9,9 @@ import { getTrackingIds } from '../shared/tracking_ids';
 import { Banner, BannerType } from './Banner';
 import BannerPresenter from '../shared/banner_presenter';
 import Translations from '../shared/messages/de';
-import DonationForm from '../shared/components/ui/form/DonationForm';
+import DonationForm from './components/ui/form/MultiStepDonationForm';
 import BannerText from './components/BannerText';
+import ButtonText from './components/ButtonTextCtrl';
 import fundsModalData from '../node_modules/fundraising-frontend-content/i18n/de_DE/data/useOfFunds.json';
 import { createCampaignProjection } from '../shared/campaign_projection';
 import { createFormItems } from './form_items';
@@ -38,6 +39,7 @@ bannerPresenter.present(
 		fundsModalData,
 		donationForm: DonationForm,
 		bannerText: BannerText,
+		buttonText: ButtonText,
 		translations: Translations,
 		formItems: createFormItems( Translations, formatters.amountInputFormatter ),
 		bannerType: BannerType.CTRL
