+++ desktop/banner_var.js
@@ -11,9 +11,10 @@ import BannerPresenter from '../shared/banner_presenter';
 import Translations from '../shared/messages/de';
 import DonationForm from './components/ui/form/MultiStepDonationForm';
 import BannerText from './components/BannerText';
+import ButtonText from './components/ButtonTextVar';
 import fundsModalData from '../node_modules/fundraising-frontend-content/i18n/de_DE/data/useOfFunds.json';
 import { createCampaignProjection } from '../shared/campaign_projection';
-import { createFormItems } from './form_items_var';
+import { createFormItems } from './form_items';
 import { LocalImpressionCount } from '../shared/local_impression_count';
 
 const bannerContainer = document.getElementById( 'WMDE-Banner-Container' );
@@ -38,6 +39,7 @@ bannerPresenter.present(
 		fundsModalData,
 		donationForm: DonationForm,
 		bannerText: BannerText,
+		buttonText: ButtonText,
 		translations: Translations,
 		formItems: createFormItems( Translations, formatters.amountInputFormatter ),
 		bannerType: BannerType.VAR
