+++ desktop/form_items.js
@@ -1,7 +1,9 @@
-import FormItemsBuilder, { Intervals, PaymentMethods } from '../shared/components/ui/form/FormItemsBuilder';
+import { AddressType } from './components/ui/form/FormItemsBuilder';
+import { Intervals, PaymentMethods } from '../shared/components/ui/form/FormItemsBuilder';
+import LocalFormItemsBuilder from './components/ui/form/FormItemsBuilder';
 
 export function createFormItems( translations, amountFormatter ) {
-	const builder = new FormItemsBuilder( translations, amountFormatter );
+	const builder = new LocalFormItemsBuilder( translations, amountFormatter );
 	builder.setIntervals(
 		Intervals.ONCE,
 		Intervals.MONTHLY,
@@ -15,6 +17,10 @@ export function createFormItems( translations, amountFormatter ) {
 		PaymentMethods.BANK_TRANSFER,
 		PaymentMethods.CREDIT_CARD,
 		PaymentMethods.PAYPAL
+	);
+	builder.setAddressType(
+		AddressType.YES,
+		AddressType.NO
 	);
 	return builder.getItems();
 }
