+++ desktop/styles/styles_ctrl.pcss
@@ -9,4 +9,5 @@
 @import './Infobox.pcss';
 @import 'BannerText.pcss';
 @import './ProgressBar.pcss';
+@import './BackLink.pcss';
 @import '../../shared/components/ui/FundsModal.pcss';
