+++ desktop/styles/variables_ctrl.pcss
@@ -24,7 +24,7 @@ Always work 'mobile first':
 	do not use @media ( max-width: $breakpoint-x )
 	always use @media ( min-width: $breakpoint-x )
 */
-$breakpoint_s: 1024px;
+$breakpoint_s: 1090px; /* TODO: Remember to swap this back to 1024 after the C20_02 test */
 $breakpoint_m: 1090px;
 $breakpoint_l: 1300px;
 $breakpoint_xl: 1580px;
