+++ package.json
@@ -25,10 +25,10 @@
     "prop-types": "^15.7.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.5",
+    "@babel/core": "^7.11.0",
     "@babel/plugin-proposal-class-properties": "^7.10.4",
     "@babel/plugin-transform-react-jsx": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
     "@babel/register": "^7.10.5",
     "autoprefixer": "^7.2.6",
     "babel-eslint": "^10.1.0",
@@ -47,6 +47,7 @@
     "minimatch": "^3.0.4",
     "mocha": "^5.2.0",
     "mocha-jsdom": "^1.2.0",
+    "postcss-calc": "^7.0.2",
     "postcss-combine-duplicated-selectors": "^6.1.0",
     "postcss-custom-properties": "^8.0.11",
     "postcss-import": "^12.0.1",
@@ -59,7 +60,7 @@
     "sinon": "^4.5.0",
     "style-loader": "^0.23.1",
     "stylelint": "^10.1.0",
-    "stylelint-config-wikimedia": "^0.10.1",
+    "stylelint-config-wikimedia": "^0.6.0",
     "toml": "^2.3.6",
     "url": "^0.11.0",
     "webpack": "^4.44.1",
